Additional refactoring for tomodachi.Options
