+++ examples/basic_examples/aws_sns_sqs_service_localstack.py
@@ -28,7 +28,7 @@ class ExampleAWSSNSSQSService(tomodachi.Service):
             aws_access_key_id="AKIAXXXXXXXXXXXXXXXX",  # Specify AWS access key (example: "AKIA****************"")
             aws_secret_access_key="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX",  # Specify AWS secret key (example: "****************************************")
         ),
-        aws_endpoint_urls=Options.AWSEndpointUrls(
+        aws_endpoint_urls=Options.AWSEndpointURLs(
             sns="http://localhost:4567",  # For example 'http://localhost:4566' (or 4567, port may vary) if localstack is used for testing
             sqs="http://localhost:4567",  # For example 'http://localhost:4566' (or 4567, port may vary) if localstack is used for testing
         ),
